requirements.txt
rumps
pynput
pyaudio
numpy
requests
pyperclip
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pywebview
py2app
